test: 로직 이관하면서 테스트 코드도 함께 이관
